+++ android/src/main/kotlin/com/newrelic/newrelic_mobile/NewrelicMobilePlugin.kt
@@ -7,11 +7,15 @@ package com.newrelic.newrelic_mobile
 
 import android.content.Context
 import android.os.Build
+import android.util.Log
 import androidx.annotation.NonNull
 import com.newrelic.agent.android.ApplicationFramework
 import com.newrelic.agent.android.FeatureFlag
 import com.newrelic.agent.android.HttpHeaders
 import com.newrelic.agent.android.NewRelic
+import com.newrelic.agent.android.logging.AgentLog
+import com.newrelic.agent.android.logging.LogLevel
+import com.newrelic.agent.android.logging.LogReporting
 import com.newrelic.agent.android.metric.MetricUnit
 import com.newrelic.agent.android.stats.StatsEngine
 import com.newrelic.agent.android.util.NetworkFailure
@@ -42,11 +46,14 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
             "getPlatformVersion" -> {
                 result.success("Android ${Build.VERSION.RELEASE}")
             }
+
             "startAgent" -> {
 
                 val applicationToken: String? = call.argument("applicationToken")
                 val dartVersion: String? = call.argument("dartVersion")
                 val loggingEnabled: Boolean? = call.argument("loggingEnabled")
+                val logLevel: String? = call.argument("logLevel")
+
 
 
                 if (call.argument<Boolean>("analyticsEventEnabled") as Boolean) {
@@ -95,21 +102,34 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 } else {
                     NewRelic.disableFeature(FeatureFlag.OfflineStorage)
                 }
+                if (call.argument<Boolean>("distributedTracingEnabled") as Boolean) {
+                    NewRelic.enableFeature(FeatureFlag.DistributedTracing)
+                } else {
+                    NewRelic.disableFeature(FeatureFlag.DistributedTracing)
+                }
+
+                if (call.argument<Boolean>("backgroundReportingEnabled") as Boolean) {
+                    NewRelic.enableFeature(FeatureFlag.BackgroundReporting)
+                } else {
+                    NewRelic.disableFeature(FeatureFlag.BackgroundReporting)
+                }
+
 
                 NewRelic.withApplicationToken(
                     applicationToken
-                ).withLoggingEnabled(loggingEnabled!!)
-                    .withLogLevel(5)
-                    .withApplicationFramework(ApplicationFramework.Flutter, "1.0.9").start(context)
+                ).withLoggingEnabled(loggingEnabled!!).withLogLevel(AgentLog.AUDIT)
+                    .withApplicationFramework(ApplicationFramework.Flutter, "1.1.0").start(context)
                 NewRelic.setAttribute("DartVersion", dartVersion)
-                StatsEngine.get().inc("Supportability/Mobile/Android/Flutter/Agent/1.0.9");
+                StatsEngine.get().inc("Supportability/Mobile/Android/Flutter/Agent/1.1.0")
                 result.success("Agent Started")
             }
+
             "setUserId" -> {
                 val userId: String? = call.argument("userId")
                 val userIsSet = NewRelic.setUserId(userId)
                 result.success(userIsSet)
             }
+
             "setAttribute" -> {
                 val name: String? = call.argument("name")
                 val value: Any? = call.argument("value")
@@ -119,27 +139,32 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                     is String -> {
                         attributeIsSet = NewRelic.setAttribute(name, value)
                     }
+
                     is Double -> {
                         attributeIsSet = NewRelic.setAttribute(name, value)
                     }
+
                     is Boolean -> {
                         attributeIsSet = NewRelic.setAttribute(name, value)
                     }
                 }
                 result.success(attributeIsSet)
             }
+
             "removeAttribute" -> {
                 val name: String? = call.argument("name")
                 val attributeIsRemoved = NewRelic.removeAttribute(name)
                 result.success(attributeIsRemoved)
             }
+
             "recordBreadcrumb" -> {
                 val name: String? = call.argument("name")
                 val eventAttributes: HashMap<String, Any>? = call.argument("eventAttributes")
 
-                val eventRecorded = NewRelic.recordBreadcrumb(name, eventAttributes);
+                val eventRecorded = NewRelic.recordBreadcrumb(name, eventAttributes)
                 result.success(eventRecorded)
             }
+
             "recordCustomEvent" -> {
                 val eventType: String? = call.argument("eventType")
                 val eventName: String? = call.argument("eventName")
@@ -150,11 +175,11 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                     result.success(eventRecorded)
                 } else {
                     val copyOfEventAttributes = eventAttributes.clone() as HashMap<*, *>
-                    for (key in  copyOfEventAttributes.keys)  {
+                    for (key in copyOfEventAttributes.keys) {
                         val value = copyOfEventAttributes[key]
-                        if(value is HashMap<*, *>) {
+                        if (value is HashMap<*, *>) {
                             for (k in value.keys) {
-                                value[k]?.let { eventAttributes.put(k as String, it) };
+                                value[k]?.let { eventAttributes.put(k as String, it) }
                             }
                             eventAttributes.remove(key)
                         }
@@ -164,24 +189,28 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                     result.success(eventRecorded)
                 }
             }
+
             "startInteraction" -> {
                 val actionName: String? = call.argument("actionName")
 
-                val interactionId = NewRelic.startInteraction(actionName);
+                val interactionId = NewRelic.startInteraction(actionName)
                 result.success(interactionId)
             }
+
             "endInteraction" -> {
                 val interactionId: String? = call.argument("interactionId")
 
                 NewRelic.endInteraction(interactionId)
                 result.success("interaction Ended")
             }
+
             "setInteractionName" -> {
                 val interactionName: String? = call.argument("interactionName")
 
                 NewRelic.setInteractionName(interactionName)
                 result.success("interaction Recorded")
             }
+
             "recordError" -> {
 
                 val exceptionMessage: String? = call.argument("exception")
@@ -214,6 +243,7 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 val bool = NewRelic.recordHandledException(exception, exceptionAttributes)
                 result.success(bool)
             }
+
             "noticeHttpTransaction" -> {
 
                 val url: String = call.argument("url")!!
@@ -244,38 +274,42 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 result.success("Http Transaction Recorded")
 
             }
+
             "noticeNetworkFailure" -> {
 
                 val url: String = call.argument("url")!!
                 val httpMethod: String = call.argument("httpMethod")!!
                 val startTime: Long = call.argument("startTime")!!
                 val endTime: Long = call.argument("endTime")!!
-                val errorCode:Int = call.argument("errorCode")!!
+                val errorCode: Int = call.argument("errorCode")!!
 
-                val nf = NetworkFailure.fromErrorCode(errorCode);
+                val nf = NetworkFailure.fromErrorCode(errorCode)
 
                 NewRelic.noticeNetworkFailure(
                     url,
                     httpMethod,
                     startTime,
                     endTime,
-                    nf)
+                    nf
+                )
                 result.success("Network Failure Recorded")
 
             }
+
             "noticeDistributedTrace" -> {
 
-                val traceContext = NewRelic.noticeDistributedTrace(null);
+                val traceContext = NewRelic.noticeDistributedTrace(null)
 
-                val traceAttributes = HashMap<String, Any>();
+                val traceAttributes = HashMap<String, Any>()
 
-                traceAttributes.putAll(traceContext.asTraceAttributes());
+                traceAttributes.putAll(traceContext.asTraceAttributes())
 
                 for (header in traceContext.headers) {
-                    traceAttributes[header.headerName] = header.headerValue;
+                    traceAttributes[header.headerName] = header.headerValue
                 }
-                result.success(traceAttributes);
+                result.success(traceAttributes)
             }
+
             "setMaxEventBufferTime" -> {
                 val maxBufferTimeInSec: Int? = call.argument("maxBufferTimeInSec")
 
@@ -284,6 +318,7 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 }
                 result.success("MaxEvent BufferTime set")
             }
+
             "setMaxEventPoolSize" -> {
                 val maxSize: Int? = call.argument("maxSize")
 
@@ -293,6 +328,7 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 result.success("maxSize set")
 
             }
+
             "setMaxOfflineStorageSize" -> {
                 val megaBytes: Int? = call.argument("megaBytes")
 
@@ -308,7 +344,7 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 val value: Double? = call.argument("value")
 
                 val isIncreased: Boolean = if (value != null) {
-                    NewRelic.incrementAttribute(name,value)
+                    NewRelic.incrementAttribute(name, value)
                 } else {
                     NewRelic.incrementAttribute(name)
                 }
@@ -325,11 +361,11 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
 
 
                 value?.let {
-                            NewRelic.recordMetric(name,category,
-                                0, it,0.0,
-                                countUnit?.let { it2 -> MetricUnit.valueOf(it2) },
-                                valueUnit?.let { it3 -> MetricUnit.valueOf(it3) })
-                };
+                    NewRelic.recordMetric(name, category,
+                        0, it, 0.0,
+                        countUnit?.let { it2 -> MetricUnit.valueOf(it2) },
+                        valueUnit?.let { it3 -> MetricUnit.valueOf(it3) })
+                }
                 result.success("Recorded Metric")
 
             }
@@ -338,16 +374,25 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
                 NewRelic.shutdown()
                 result.success("agent is shutDown")
             }
+
             "currentSessionId" -> {
                 result.success(NewRelic.currentSessionId())
             }
+
             "addHTTPHeadersTrackingFor" -> {
                 val headers: ArrayList<String>? = call.argument("headers") as ArrayList<String>?
                 result.success(NewRelic.addHTTPHeadersTrackingFor(headers))
             }
+
             "getHTTPHeadersTrackingFor" -> {
                 result.success(HttpHeaders.getInstance().httpHeaders.toList())
             }
+            "logAttributes" -> {
+                val attributes: HashMap<String, Any>? = call.argument("attributes")
+                NewRelic.logAttributes(attributes)
+                result.success("Recorded Log")
+            }
+
             else -> {
                 result.notImplemented()
             }
@@ -375,6 +420,7 @@ class NewrelicMobilePlugin : FlutterPlugin, MethodCallHandler {
 
     override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
         channel.setMethodCallHandler(null)
+
     }
 }