+++ example/lib/main.dart
@@ -11,6 +11,7 @@ import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
+import 'package:newrelic_mobile/loglevel.dart';
 import 'package:newrelic_mobile/metricunit.dart';
 import 'package:newrelic_mobile/config.dart';
 import 'package:newrelic_mobile/network_failure.dart';
@@ -52,10 +53,11 @@ void main() {
       webViewInstrumentation: true,
       printStatementAsEventsEnabled: true,
       httpInstrumentationEnabled: true,
+      distributedTracingEnabled: true,
       fedRampEnabled: false);
 
   NewrelicMobile.instance.start(config, () {
-    runApp(MyApp());
+    runApp(const MyApp());
   });
   NewrelicMobile.instance.setMaxEventPoolSize(3000);
   NewrelicMobile.instance.setMaxEventBufferTime(200);
@@ -66,17 +68,17 @@ void main() {
 /// The main app.
 class MyApp extends StatelessWidget {
   /// Creates an [App].
-  MyApp({Key? key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       navigatorObservers: [NewRelicNavigationObserver()],
       routes: {
-        'pageone': (context) => Page1Screen(),
+        'pageone': (context) => const Page1Screen(),
         'pagetwo': (context) => Page2Screen(),
-        'pagethree': (context) => Page3Screen(),
-        'pagefour': (context) => Page4Screen()
+        'pagethree': (context) => const Page3Screen(),
+        'pagefour': (context) => const Page4Screen()
       },
       theme: ThemeData(
         primarySwatch: Colors.blue,
@@ -103,7 +105,52 @@ class Page1Screen extends StatelessWidget {
               children: <Widget>[
                 ElevatedButton(
                     onPressed: () async {
-                      print(NewrelicMobile.instance.currentSessionId());
+                      NewrelicMobile.instance.logInfo("testing logs");
+                      NewrelicMobile.instance.logDebug("testing logs debug");
+                      NewrelicMobile.instance
+                          .logWarning("testing logs warning");
+                      NewrelicMobile.instance
+                          .logVerbose("testing logs verbose");
+                      NewrelicMobile.instance
+                          .log(LogLevel.ERROR, "testing logs");
+                      NewrelicMobile.instance.logInfo("testing logs");
+                      NewrelicMobile.instance.logInfo("testing logs");
+
+                      var map = {};
+                      map["test12"] = "value";
+                      map["test23"] = "value1";
+                      map["logLevel"] = "INFO";
+                      map["message"] = "testing logs with attributes";
+
+                      NewrelicMobile.instance
+                          .logAttributes(map.cast<String, dynamic>());
+
+                      NewrelicMobile.instance
+                          .logAll(Exception("This is an exception"), {
+                        "BreadNumValue": 12.3,
+                        "BreadStrValue": "FlutterBread",
+                        "BreadBoolValue": true,
+                        "message": "This is a message with attributes"
+                      });
+
+                      NewrelicMobile.instance.logAttributes({
+                        "BreadNumValue": 12.3,
+                        "BreadStrValue": "FlutterBread",
+                        "BreadBoolValue": true,
+                        "message": "This is a message with attributes"
+                      });
+
+                      if (kDebugMode) {
+                        print(await NewrelicMobile.instance.currentSessionId());
+                      }
+                      // showDialog<String>(
+                      //   barrierDismissible: false,
+                      //   context: context,
+                      //   builder: (BuildContext context) => const PopPopPop(),
+                      // );
+                      if (kDebugMode) {
+                        print(NewrelicMobile.instance.currentSessionId());
+                      }
                       NewrelicMobile.instance.incrementAttribute(
                           "FlutterCustomAttrNumber",
                           value: 5.0);
@@ -131,7 +178,9 @@ class Page1Screen extends StatelessWidget {
                       final response = await request.close();
 
                       response.transform(utf8.decoder).listen((contents) {
-                        print(contents);
+                        if (kDebugMode) {
+                          print(contents);
+                        }
                       });
                     },
                     child: const Text('Http call to non-existing URL',
@@ -151,7 +200,9 @@ class Page1Screen extends StatelessWidget {
                       final response = await request.close();
 
                       response.transform(utf8.decoder).listen((contents) {
-                        print(contents);
+                        if (kDebugMode) {
+                          print(contents);
+                        }
                       });
                     },
                     child: const Text('Http Default Client',
@@ -169,7 +220,9 @@ class Page1Screen extends StatelessWidget {
                       //     'http://graph.facebook.com/');
                       // var response = await http.get(url);
                       // print('Response status: ${await response.stream.bytesToString()}');
-                      print('Response body: ${response.statusCode}');
+                      if (kDebugMode) {
+                        print('Response body: ${response.statusCode}');
+                      }
                     },
                     child: const Text('Http Library ',
                         maxLines: 1, textDirection: TextDirection.ltr)),
@@ -181,9 +234,13 @@ class Page1Screen extends StatelessWidget {
                         dio.options.followRedirects = false;
                         var response =
                             await dio.get('http://graph.facebook.com/');
-                        print(response);
+                        if (kDebugMode) {
+                          print(response);
+                        }
                       } catch (e) {
-                        print(e);
+                        if (kDebugMode) {
+                          print(e);
+                        }
                       }
                     },
                     child: const Text('Http Dio Library ',
@@ -194,9 +251,13 @@ class Page1Screen extends StatelessWidget {
                         var dio = Dio();
                         var response = await dio.get(
                             'https://cb6b02be-a319-4de5-a3b1-361de2564493.mock.pstmn.io/searchpage');
-                        print(response);
+                        if (kDebugMode) {
+                          print(response);
+                        }
                       } catch (e) {
-                        print(e);
+                        if (kDebugMode) {
+                          print(e);
+                        }
                       }
                     },
                     child: const Text('OOM Issue Library ',
@@ -219,9 +280,13 @@ class Page1Screen extends StatelessWidget {
                         var response = await dio.post(
                             'https://reqres.in/api/register',
                             data: "{ 'email': 'sydney@fife'}");
-                        print(response.data);
+                        if (kDebugMode) {
+                          print(response.data);
+                        }
                       } catch (e) {
-                        print(e);
+                        if (kDebugMode) {
+                          print(e);
+                        }
                       }
                     },
                     child: const Text('Http Dio Post Library ',
@@ -229,7 +294,9 @@ class Page1Screen extends StatelessWidget {
                 Query(
                     options: QueryOptions(document: gql(readCounters)),
                     builder: (result, {fetchMore, refetch}) {
-                      print(result.data.toString());
+                      if (kDebugMode) {
+                        print(result.data.toString());
+                      }
                       // If stements here to check handle different states;
                       if (result.isLoading) {
                         return const Center(
@@ -262,7 +329,7 @@ class Page1Screen extends StatelessWidget {
 class Page2Screen extends StatelessWidget {
   /// Creates a [Page2Screen].
 
-  var interActionId;
+  dynamic interActionId;
 
   Page2Screen({Key? key, this.interActionId}) : super(key: key);
 
@@ -296,7 +363,7 @@ class Page2Screen extends StatelessWidget {
               },
               child: const Text('State Error'),
             ),
-            Row(
+            const Row(
               children: [
                 Text(
                   "ErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorErrorError",
@@ -306,7 +373,9 @@ class Page2Screen extends StatelessWidget {
             ),
             ElevatedButton(
               onPressed: () {
-                print("test");
+                if (kDebugMode) {
+                  print("test");
+                }
                 debugPrint("test");
                 throw TypeError();
               },
@@ -356,7 +425,7 @@ class Page3Screen extends StatelessWidget {
             children: <Widget>[
               ElevatedButton(
                 onPressed: () {
-                  for (var i = 0; i < 100; i++)
+                  for (var i = 0; i < 100; i++) {
                     NewrelicMobile.instance.recordCustomEvent(
                         "Test Custom Event",
                         eventName: "User Purchase",
@@ -365,6 +434,7 @@ class Page3Screen extends StatelessWidget {
                           "price": 34.00,
                           "loop test": i
                         });
+                  }
                 },
                 child: const Text('Record Custom Event'),
               ),
@@ -381,10 +451,14 @@ class Page3Screen extends StatelessWidget {
                     var dio = Dio();
                     var response =
                         await dio.get('https://reqres.in/api/users?delay=15');
-                    print(response);
+                    if (kDebugMode) {
+                      print(response);
+                    }
                     NewrelicMobile.instance.endInteraction(id);
                   } catch (e) {
-                    print(e);
+                    if (kDebugMode) {
+                      print(e);
+                    }
                   }
                 },
                 child: const Text('Interaction Example'),
@@ -425,12 +499,24 @@ class _Page4ScreenState extends State<Page4Screen> {
                 style: Theme.of(context).textTheme.titleLarge,
               ),
               ElevatedButton(
-                onPressed: () {
-                  computeService.fetchUser().then((value) {
-                    setState(() {
-                      person = value;
-                    });
+                onPressed: () async {
+                  ReceivePort port = ReceivePort();
+                  var errorPort = ReceivePort();
+                  errorPort.listen((message) {
+                    if (kDebugMode) {
+                      print('Error: $message');
+                    }
+                    NewrelicMobile.instance
+                        .recordError(message, StackTrace.current);
                   });
+                  await Isolate.spawn(_isolateFunction, port.sendPort,
+                      onError: errorPort.sendPort);
+
+                  // computeService.fetchUser().then((value) {
+                  //   setState(() {
+                  //     person = value;
+                  //   });
+                  // });
                 },
                 child: const Text('Isolate Compute Error'),
               ),
@@ -456,6 +542,10 @@ class _Page4ScreenState extends State<Page4Screen> {
       );
 }
 
+void _isolateFunction(_) {
+  throw Exception('Uncaught error in isolate');
+}
+
 class ComputeService {
   Future<Person> fetchUser() async {
     String userData = await Api.getUser("Compute");
@@ -463,7 +553,7 @@ class ComputeService {
   }
 
   Person deserializeJson(String data) {
-    throw new Error();
+    throw Error();
   }
 }
 
@@ -491,8 +581,10 @@ class SpawnService {
   void deserializePerson(List<dynamic> values) async {
     var dio = Dio();
     var response = await dio.get('https://reqres.in/api/users?delay=15');
-    print(response);
-    throw new Exception("this is isplation error");
+    if (kDebugMode) {
+      print(response);
+    }
+    throw Exception("this is isplation error");
   }
 }
 
@@ -519,3 +611,40 @@ const rickCharacters = '''
     }
   }
  ''';
+
+class PopPopPop extends StatelessWidget {
+  const PopPopPop({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+        title: const Text('Here we go...'),
+        icon: const Icon(
+          Icons.error_outline,
+          color: Colors.red,
+          size: 18.0,
+        ),
+        content: const SingleChildScrollView(
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text('Something went wrong but we\'re on it.'),
+            ],
+          ),
+        ),
+        actions: <Widget>[
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: [
+              TextButton(
+                onPressed: () {
+                  Navigator.pushNamed(context, 'pagetwo',
+                      arguments: {'id': ""});
+                },
+                child: const Text('OK'),
+              ),
+            ],
+          )
+        ]);
+  }
+}