+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 1.1.0
+## New Features
+
+1. Application Exit Information
+    - Added ApplicationExitInfo to data reporting
+    - Enabled by default
+
+2. Log Forwarding to New Relic
+    - Implement static API for sending logs to New Relic
+    - Can be enabled/disabled in your mobile application's entity settings page
+   
+3. A new distributedTracingEnabled feature flag has been introduced to enable disable distributed Tracing functionality.
+
+
+## Improvements
+
+- Native Android agent updated to version 7.5.0
+- Native iOS agent updated to version 7.5.0
+
 ## 1.0.9
 * Improvements